[Refactor] refactor homePage codes
