+++ src/components/Home/CustomTopic.tsx
@@ -1,0 +1,106 @@
+export function CustomTopic() {
+  return (
+    <section>
+      <div className='jb flex w-screen flex-col items-center px-[12px]  pt-20 pb-8 text-center lg:pt-[100px] lg:pb-[124px]'>
+        <h2>客製化諮商主題</h2>
+        <p className='subTitle'>
+          Lorem ipsum dolor sit amet consectetur adipiscing eli mattis sit
+          phasellus mollis sit aliquam sit nullam.
+        </p>
+
+        {/* 卡片 */}
+        <ul className='mt-[34px] flex w-[366px] flex-wrap items-center justify-around lg:w-[1012px]'>
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+
+          <li className='mb-10 w-[180px] lg:w-[316px]'>
+            {/* 這是圖片 */}
+            <div className='h-[180px] w-[180px] rounded-t-2xl border border-b-0 border-secondary bg-secondary lg:h-[286px] lg:w-[316px]'></div>
+
+            {/* 文字區塊 */}
+            <div className='rounded-b-2xl border border-t-0 border-secondary pt-4 pb-7 lg:pt-5 lg:pb-8'>
+              <h3 className='mb-1 font-bold text-primary-heavy lg:text-xl'>
+                女性議題
+              </h3>
+              <p className='text-xs text-[#9795B5] lg:text-base'>
+                產後憂鬱、更年期
+              </p>
+            </div>
+          </li>
+        </ul>
+      </div>
+    </section>
+  );
+}
