+++ src/lib/homeFilesRoute.ts
@@ -1,0 +1,24 @@
+import { Banner } from '../components/Home/Banner'
+import { SuggestCounselor } from '../components/Home/SuggestCounselor'
+import { CustomTopic } from '../components/Home/CustomTopic'
+import { PlatformFeature } from '../components/Home/PlatformFeature'
+import { UserComment } from '../components/Home/UserComment'
+import { ReservationTour } from '../components/Home/ReservationTour'
+
+export const homeComponents = {
+  Banner,
+  SuggestCounselor,
+  CustomTopic,
+  PlatformFeature,
+  UserComment,
+  ReservationTour
+}
+
+export const fakeCounselorAry = [
+  '#女性議題',
+  '#親密關係',
+  '#青少年',
+  '#中老年議題',
+  '#一般成人',
+  '#PTSD'
+]
