+++ src/pages/_app.tsx
@@ -1,6 +1,15 @@
 import '@/styles/globals.css'
+import '../styles/antd/style.css'
 import type { AppProps } from 'next/app'
+import { Header } from '@/layout/Header'
+import { Footer } from '@/layout/Footer'
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return (
+    <>
+      <Header />
+      <Component {...pageProps} />
+      <Footer />
+    </>
+  )
 }
